+++ src/main.rs
@@ -10,16 +10,15 @@ mod widgets;
 mod account;
 mod init;
 mod instance;
-mod logs;
 mod settings;
 
 fn main() {
+    #[cfg(debug_assertions)]
     unsafe {
-        #[cfg(debug_assertions)]
         std::env::set_var("RUST_BACKTRACE", "1");
     }
 
     if let Err(e) = app::run() {
-        log::error!("{e:#?}");
+        eprintln!("{e:#?}");
     }
 }