+++ src/utils/mod.rs
@@ -11,13 +11,14 @@ pub mod message;
 pub mod sha1;
 
 pub trait ShowWindow {
-    // I'll use any for an async mutex
+    // Just a mutex or an unused var tbh
     fn show(
         &mut self,
         mctx: Context,
         ctx: &Context,
         show_win: Arc<AtomicBool>,
-        data: Arc<dyn Any + Sync + Send>,
+        data1: Arc<dyn Any + Sync + Send>,
+        data2: Arc<dyn Any + Sync + Send>,
         cl: Client,
     );
 }