+++ CHANGELOG.md
@@ -13,6 +13,8 @@
     function
 -   The native libraries aren't extracted into the minecraft\_cache directory
     anymore, it's on the instance directory now
+-   And minecraft\_cache was returned to cache
+-   Our oldest loadable version is 1.2.5 rn
 
 # Reminders (to me)
 
@@ -21,3 +23,8 @@
 -   Do something with the 'Settings' and 'Account' windows
 -   Maybe add an optional thread handle to the instances to see if it's
     currently running or not so that it can be stopped by the gui
+-   Figure out why the instance icons are not on the same line
+-   Nvm the deferred viewports are working correctly, it's just blocking when
+    the download happens as it tries to lock the same instances mutex to get
+    the version list and the main thread getting the actual instancs
+    (I even removed tokio just for me to realize this)